Recipes #8 fission recipe type

+++ src/main/java/igentuman/nc/container/NCEnergyContainer.java
@@ -169,7 +169,7 @@ public class NCEnergyContainer extends AbstractContainerMenu {
         return blockEntity.getEnergy().orElse(null);
     }
 
-    public int getProgress() {
+    public double getProgress() {
         return blockEntity.getProgress();
     }
 }
