+++ src/main/java/igentuman/nc/gui/IProgressScreen.java
@@ -1,5 +1,5 @@
 package igentuman.nc.gui;
 
 public interface IProgressScreen {
-    int getProgress();
+    double getProgress();
 }
